+++ hphp/runtime/vm/jit/unwind-x64.h
@@ -17,168 +17,18 @@
 #ifndef incl_HPHP_VM_TRANSLATOR_UNWIND_X64_H_
 #define incl_HPHP_VM_TRANSLATOR_UNWIND_X64_H_
 
-#include <cstdlib>
-#include <sstream>
-#include <string>
-#include <memory>
-#include <exception>
-#include <typeinfo>
+#include "hphp/runtime/vm/jit/unwind-types.h"
 
-#include "hphp/runtime/base/rds.h"
-#include "hphp/runtime/base/typed-value.h"
-#include "hphp/runtime/vm/jit/types.h"
-#include "hphp/runtime/vm/tread-hash-map.h"
-#include "hphp/util/asm-x64.h"
-#include "hphp/util/assertions.h"
-
-#ifndef _MSC_VER
-#include <unwind.h>
-#else
-// Stubs! Don't you just love them?
-// In this case, the stubs are just
-// here to allow the code to compile.
-// Attempting to use the stubs will
-// likely just result in a segfault.
-
-struct _Unwind_Exception {
-  uint64_t exception_class;
-};
-
-#define _URC_CONTINUE_UNWIND 0
-#define _URC_INSTALL_CONTEXT 0
-#define _URC_HANDLER_FOUND 0
-typedef int _Unwind_Reason_Code;
-
-#define _UA_HANDLER_FRAME 0
-#define _UA_CLEANUP_PHASE 0
-#define _UA_SEARCH_PHASE 0
-typedef int _Unwind_Action;
-
-typedef void _Unwind_Context;
-
-inline uintptr_t _Unwind_GetGR(_Unwind_Context*, int) {
-  always_assert(false);
-  return 0;
-}
-inline uintptr_t _Unwind_GetIP(_Unwind_Context*) {
-  always_assert(false);
-  return 0;
-}
-inline void _Unwind_Resume() { always_assert(false); }
-inline void _Unwind_SetIP(_Unwind_Context*, uint64_t) { always_assert(false); }
-#endif
-
-namespace HPHP {
-struct ActRec;
-
-namespace jit {
+namespace HPHP { namespace jit { namespace x64 {
 
 //////////////////////////////////////////////////////////////////////
 
-typedef TreadHashMap<CTCA, TCA, ctca_identity_hash> CatchTraceMap;
-
-/*
- * Information the unwinder needs stored in RDS, and the rds::Link for
- * it.  Used to pass values between unwinder code and catch traces.
- */
-struct UnwindRDS {
-  /* When a cleanup (non-side-exiting) catch trace is executing, this will
-   * point to the currently propagating exception, to be passed to
-   * _Unwind_Resume at the end of cleanup. */
-  _Unwind_Exception* exn;
-
-  /* Some helpers need to signal an error along with a TypedValue to be pushed
-   * on the eval stack. When present, that value lives here. */
-  TypedValue unwinderTv;
-
-  /* When returning from a frame that had its m_savedRip smashed by the
-   * debugger, the return stub stashes values here to be used after running the
-   * appropriate catch trace. In addition, a non-nullptr debuggerReturnSP is
-   * used as the flag to endCatchHelper that it should perform a
-   * REQ_POST_DEBUGGER_RET rather than resuming the unwind process. */
-  TypedValue* debuggerReturnSP;
-  Offset debuggerReturnOff;
-
-  /* This will be true iff the currently executing catch trace should side exit
-   * to somewhere else in the TC, rather than resuming the unwind process. */
-  bool doSideExit;
-};
-extern rds::Link<UnwindRDS> unwindRdsInfo;
-
-inline ptrdiff_t unwinderExnOff() {
-  return unwindRdsInfo.handle() + offsetof(UnwindRDS, exn);
-}
-
-inline ptrdiff_t unwinderSideExitOff() {
-  return unwindRdsInfo.handle() + offsetof(UnwindRDS, doSideExit);
-}
-
-inline ptrdiff_t unwinderTvOff() {
-  return unwindRdsInfo.handle() + offsetof(UnwindRDS, unwinderTv);
-}
-
-inline ptrdiff_t unwinderDebuggerReturnOffOff() {
-  return unwindRdsInfo.handle() + offsetof(UnwindRDS, debuggerReturnOff);
-}
-
-inline ptrdiff_t unwinderDebuggerReturnSPOff() {
-  return unwindRdsInfo.handle() + offsetof(UnwindRDS, debuggerReturnSP);
-}
-
-//////////////////////////////////////////////////////////////////////
-
-/*
- * Meant to work like __cxxabiv1::__is_dependent_exception
- * See libstdc++-v3/libsupc++/unwind-cxx.h in GCC
- */
-static inline bool isDependentException(uint64_t c)
-{
-  return (c & 1);
-}
-
-/**
- * Meant to work like __cxxabiv1::__get_object_from_ue() but with a specific
- * return type.
- * See libstdc++-v3/libsupc++/unwind-cxx.h in GCC
- */
-inline std::exception* exceptionFromUnwindException(
-  _Unwind_Exception* exceptionObj)
-{
-  constexpr size_t sizeOfDependentException = 112;
-  if (isDependentException(exceptionObj->exception_class)) {
-    return *reinterpret_cast<std::exception**>(
-      reinterpret_cast<char*>(exceptionObj + 1) - sizeOfDependentException);
-  } else {
-    return reinterpret_cast<std::exception*>(exceptionObj + 1);
-  }
-}
-
-inline const std::type_info& typeInfoFromUnwindException(
-  _Unwind_Exception* exceptionObj
-  )
-{
-  if (isDependentException(exceptionObj->exception_class)) {
-    // like __cxxabiv1::__get_refcounted_exception_header_from_obj()
-    constexpr size_t sizeOfRefcountedException = 128;
-    char * obj = reinterpret_cast<char*>(
-        exceptionFromUnwindException(exceptionObj));
-    char * header = obj - sizeOfRefcountedException;
-    // Dereference the exc field, the type_info* is the first field inside that
-    constexpr size_t excOffset = 16;
-    return *reinterpret_cast<std::type_info*>(header + excOffset);
-  } else {
-    // like __cxxabiv1::__get_exception_header_from_ue()
-    constexpr size_t sizeOfCxaException = 112;
-    return **reinterpret_cast<std::type_info**>(
-      reinterpret_cast<char*>(exceptionObj + 1) - sizeOfCxaException);
-  }
-}
+TCA lookup_catch_trace(TCA rip, _Unwind_Exception* exn);
 
 /*
  * Called whenever we create a new translation cache for the whole
  * region of code.
  */
-typedef std::shared_ptr<void> UnwindInfoHandle;
 UnwindInfoHandle register_unwind_region(unsigned char* address, size_t size);
 
 /*
@@ -191,14 +41,10 @@ tc_unwind_personality(int version,
                       _Unwind_Exception* exceptionObj,
                       _Unwind_Context* context);
 
-struct TCUnwindInfo {
-  TCA catchTrace;
-  ActRec* fp;
-};
 TCUnwindInfo tc_unwind_resume(ActRec* fp);
 
 //////////////////////////////////////////////////////////////////////
 
-}}
+}}}
 
 #endif
