+++ hphp/runtime/vm/unwind-vm.h
@@ -13,8 +13,8 @@
    | license@php.net so we can mail you a copy immediately.               |
    +----------------------------------------------------------------------+
 */
-#ifndef incl_HPHP_UNWIND_H_
-#define incl_HPHP_UNWIND_H_
+#ifndef incl_HPHP_UNWIND_VM_H_
+#define incl_HPHP_UNWIND_VM_H_
 
 #include <stdexcept>
 #include "hphp/runtime/base/exceptions.h"
@@ -103,8 +103,8 @@ struct VMSwitchModeBuiltin : std::exception {
 
 }
 
-#define incl_HPHP_VM_UNWIND_INL_H_
-#include "hphp/runtime/vm/unwind-inl.h"
-#undef incl_HPHP_VM_UNWIND_INL_H_
+#define incl_HPHP_VM_UNWIND_VM_INL_H_
+#include "hphp/runtime/vm/unwind-vm-inl.h"
+#undef incl_HPHP_VM_UNWIND_VM_INL_H_
 
 #endif
